Write SessionsController#create and associated tests
redirect to newly-logged-in user; invalid credentials cause the
new session template to be rendered, flash errors set, and 401 status
code applied.

SessionsController#destroy is also included in this commit; it will
call ApplicationController.logout!, which has yet to be written. At this
time, the #destroy action only serves to render the new session template.