Update SessionsController#create to set flash[:alert]
Previously, flash[:errors] was set in the case of a failed
login attempt. This commit changes this behavior in two ways. First,
the flash is replaced with flash.now, to reflect that the controller
will render (rather than redirect to) new_session_url. Second, instead
of setting :errors, we set :alert.

Tests in sessions_controller_spec are also changed to reflect this
new behavior of the controller.