Add feature testing for failed sign in scenarios
This commit builds on the framework from a89ee230b9aa58425996d0bfaad90d5c55457660
by adding tests for the three username/password scenarios in which
login credentials will be invalid.

To this end, this commit includes a new shared example to test for
proper rendering, error (alert) handling, and http status in this
situation.